requirements.txt
black==24.10.0
django-allauth==65.1.0
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
gunicorn==23.0.0
importlib-metadata==8.0.0
inflect==7.3.1
ipython==8.12.3
jaraco.collections==5.1.0
nbconvert==7.16.5
pandas==2.2.3
pip-chill==1.0.3
psycopg2-binary==2.9.10
python-binance==1.0.27
python-dotenv==1.0.1
sshtunnel==0.4.0
tinycss2==1.4.0
tomli==2.0.1
yarg==0.1.9
